Restructured fs to organize product details
