--- components/Middle.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Middle = () => {
-  return <div>Im going to build the middle</div>;
-};
-
-export default Middle;
