+++ components/ProductInfo/Cart.js
@@ -1,5 +1,5 @@
 import React from "react";
-import style from "../styles/Cart.module.css";
+import style from "../../styles/Cart.module.css";
 
 const Cart = () => {
