+++ styles/Cart.module.css
@@ -3,6 +3,5 @@
     border-radius: 4%;
     border: 1px solid rgb(219, 216, 216);
     display: flex;
-    float: right;
 }
