+++ styles/ProdImage.module.css
@@ -27,7 +27,6 @@
   margin-right: 0.5rem;
 }
 .container {
-
   display: flex;
 }
 .share {
