+++ styles/ProdInfo.module.css
@@ -1,0 +1,5 @@
+.container{
+    display: flex;
+    flex-wrap: nowrap;
+    justify-content: space-between;
+}
